+++ tests/policy_definition/network_security/snet_05_denied_namingconform_no_nsg_attached_dedicated/terraform.tf
@@ -4,12 +4,14 @@ terraform {
       source  = "hashicorp/azurerm"
       version = ">=3.0.0"
     }
+    azapi = {
+      source  = "Azure/azapi"
+      version = "1.13.1"
+    }
   }
 }
 
 provider "azurerm" {
-  features {
-
-  }
+  features {}
   skip_provider_registration = true
 }